pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ftw-hub"
dynamic = ["version"]
description = 'Convert folk dance event data for different platforms'
readme = "README.md"
requires-python = ">=3.8"
license = "gpl-3.0-or-later"
keywords = []
authors = [
  { name = "Roman Zimmermann", email = "roman@more-onion.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click>=8.1,<9",
  "python-dateutil",
  "pyyaml",
]

[project.optional-dependencies]
dev = [
  "black",
  "ipython",
  "pre-commit",
  "pytest",
  "pytest-cov",
]

[project.urls]
Documentation = "https://github.com/torotil/ftw-hub#readme"
Issues = "https://github.com/torotil/ftw-hub/issues"
Source = "https://github.com/torotil/ftw-hub"

[project.scripts]
ftw-hub = "ftw_hub:cli"

[tool.hatch.version]
source = "vcs"

[tool.hatch.envs.default]
path = ".hatch"
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/ftw_hub tests}"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
ignore = [
  "RUF002", # Allow typographic right-ticks.
]

[tool.ruff.lint]
select = ["I"]

[tool.coverage.run]
source_pkgs = ["ftw_hub", "tests"]
branch = true
parallel = true
omit = [
  "src/ftw_hub/__about__.py",
]

[tool.coverage.paths]
ftw_hub = ["src/ftw_hub", "*/ftw-hub/src/ftw_hub"]
tests = ["tests", "*/ftw-hub/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
